Chore(CI): Correct variable name for universal APK in Telegram workflow
This commit fixes a variable name in the `.github/workflows/android.yml` file.

In the final step of the workflow, which sends the build artifact to Telegram, the `document` parameter was incorrectly referencing `${{ env.APK_FILE_ARMALL }}`. This has been corrected to `${{ env.APK_FILE_ALL }}` to accurately point to the universal APK file.